+++ src/restfw_admin/restfw_admin/typing.py
@@ -3,13 +3,10 @@
 :Authors: cykooz
 :Date: 25.04.2020
 """
-from typing import Any, Callable, Dict, List, Tuple, Union
+from typing import Any, Callable
 
 import colander
 
 
 ColanderNode = colander._SchemaNode
 ColanderValidator = Callable[[ColanderNode, Any], None]
-JsonNumber = Union[int, float]
-SimpleJsonValue = Union[str, JsonNumber, bool, None]
-Json = Union[SimpleJsonValue, List['Json'], Tuple['Json', ...], Dict[str, 'Json']]